+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include <chrono>
+#include <fstream>
 #include <iostream>
 #include <thread>
 
@@ -130,30 +131,65 @@ void mineParallel(const Settings& settings,
 }
 
 void boinc(Settings settings, const std::vector<std::string>& args) {
-  // check setup
+  // get project dir
+  auto project_env = std::getenv("PROJECT_DIR");
+  if (!project_env) {
+    Log::get().error("PROJECT_DIR environment variable not set", true);
+  }
+  std::string project_dir(project_env);
+  ensureTrailingSlash(project_dir);
 
-  // TODO: use a fixed loda home inside the users home dir
-  // TODO: make sure to lock the folder when cloning!
-  // TODO: read <user_name>...</user_name> from $PROJECT_DIR/init_data.xml
+  // set loda home dir
+  auto slots_pos = project_dir.find("slots");
+  if (slots_pos == std::string::npos) {
+    Log::get().error("Missing slots in project dir: " + project_dir, true);
+  }
+  const std::string loda_home = project_dir.substr(0, slots_pos) + "projects" +
+                                FILE_SEP + "boinc.loda-lang.org_loda";
+  Setup::setLodaHome(loda_home);
 
+  // mark setup as loaded
+  Setup::getMiningMode();
+
+  // ensure git is installed
   if (!hasGit()) {
     Log::get().error("Git not found. Please install it and try again", true);
   }
-  Setup::getLodaHome();
+
+  // clone programs repository if necessary
   if (!Setup::existsProgramsHome()) {
+    FolderLock lock(project_dir);
     if (!Setup::cloneProgramsHome()) {
       Log::get().error("Cannot clone programs repository", true);
     }
   }
+
   // configure setup
   Setup::setMiningMode(MINING_MODE_CLIENT);
   Setup::forceCPUHours();
-  Setup::setSubmittedBy("BOINC");
+  std::ifstream init_data(project_dir + "init_data.xml");
+  std::string line;
+  std::string submitted_by = "BOINC";
+  while (std::getline(init_data, line)) {
+    auto start = line.find("<user_name>");
+    auto end = line.find("</user_name>");
+    if (start != std::string::npos && end != std::string::npos) {
+      line = line.substr(start + 11);
+      line = line.substr(0, end - start - 11);
+      submitted_by = line;
+      Log::get().info("Submitting programs as " + submitted_by);
+      break;
+    }
+  }
+  Setup::setSubmittedBy(submitted_by);
+  return;
+
   // pick a random miner profile if not mining in parallel
   if (!settings.parallel_mining || settings.num_miner_instances == 1) {
     settings.miner_profile = std::to_string(Random::get().gen() % 100);
   }
-  // start mining
+
+  // start mining!
   Commands commands(settings);
   commands.mine();
 }