+++ src/include/miner.hpp
@@ -30,6 +30,10 @@ class Miner {
 
   void submit(const std::string &path, std::string id);
 
+  void reportProgress();
+
+  std::string progress_file;
+
  private:
   bool checkRegularTasks();
 
@@ -38,6 +42,10 @@ class Miner {
   void ensureSubmittedBy(Program &program);
 
   void updateSubmittedBy(Program &program);
+
+  void reportCPUHour();
+
+  int64_t getElapsedMinutes() const;
 
   static const std::string ANONYMOUS;
   static const int64_t PROGRAMS_TO_FETCH;
@@ -58,7 +66,8 @@ class Miner {
   int64_t num_new;
   int64_t num_updated;
   int64_t num_removed;
-  int64_t num_hours;
+  int64_t num_reported_hours;
   int64_t current_fetch;
   std::map<std::string, int64_t> num_received_per_profile;
+  std::chrono::steady_clock::time_point start_time;
 };